feat: custom certificate
