+++ apps/api/src/routes/api/v1/settings/types.ts
@@ -42,3 +42,8 @@ export interface DeleteSSHKey {
        id: string
     }
 }
+export interface OnlyIdInBody {
+    Body: {
+        id: string
+     }
+}
