+++ apps/ui/package.json
@@ -42,13 +42,14 @@
 	},
 	"type": "module",
 	"dependencies": {
-		"dayjs": "1.11.5",
 		"@sveltejs/adapter-static": "1.0.0-next.39",
 		"@tailwindcss/typography": "^0.5.7",
 		"cuid": "2.1.8",
 		"daisyui": "2.24.2",
+		"dayjs": "1.11.5",
 		"js-cookie": "3.0.1",
 		"p-limit": "4.0.0",
+		"svelte-file-dropzone": "^1.0.0",
 		"svelte-select": "4.4.7",
 		"sveltekit-i18n": "2.2.2"
 	}
