+++ apps/ui/src/routes/destinations/[id]/configuration/sshkey.svelte
@@ -61,7 +61,7 @@
 				<div class="pb-2 text-center font-bold">No SSH key found</div>
 				<div class="flex justify-center">
 					<a
-						href="/settings/ssh-keys"
+						href="/settings/ssh"
 						sveltekit:prefetch
 						class="add-icon bg-sky-600 hover:bg-sky-500"
 					>
