+++ others/morethanone_cert.pem
@@ -1,0 +1,33 @@
+-----BEGIN CERTIFICATE-----
+MIIFrTCCA5WgAwIBAgIUWscetQdy/o3JGjsR1LifMVHSFPYwDQYJKoZIhvcNAQEL
+BQAwZTELMAkGA1UEBhMCQVUxEzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoM
+GEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDEeMBwGA1UEAwwVbm9kZS5jb20sd3d3
+Lm5vZGUuY29tMCAXDTIyMDkyMTA5MDgzOVoYDzIwNTAwMjA2MDkwODM5WjBlMQsw
+CQYDVQQGEwJBVTETMBEGA1UECAwKU29tZS1TdGF0ZTEhMB8GA1UECgwYSW50ZXJu
+ZXQgV2lkZ2l0cyBQdHkgTHRkMR4wHAYDVQQDDBVub2RlLmNvbSx3d3cubm9kZS5j
+b20wggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQC91LQflL+X2YU1oiDl
+V9xGKtkUM7KOZgCQSkwDdkxF6fzjOAkjnekWkPRvPijWLuOLuqbMYZe8uEw06DgL
+JlwcYBKC8hrBsaT2QpbzDuetzsYeudfBVYfLSHiUOvP+pVf2nurprbHPxdWLmh8v
+HFLmTSVAH0sDDRb3gt2S7OnrGjw4fSjfS/5lFzGdV4ic1vFt2bB9Fv+D5pKPO/Zm
+O9XIhxZ8/LBNTRhR6TEWn3h0HMVqTMKVm3cjDB+xelivZ8LCV7sO2Y/ZS70ByI2K
+op4e7H50uNh4WjhSXNBSabLNw1gC/cvBLn0wypGtKgRt26Nj4FNvpfoXfbwCDkgJ
+SFyrKLaL1YraNZOxIWgcQs2dKoVYKQfLBQeL+uSe26OY5eXZUcq4cmQjzeahF0MM
+7STR6c14QuR0ppgInpuLGFVmcnCQBhpR7P2Y3ITIGqQto/ZHAWvx+BgkdUhhKoFO
+6YUvBiG14CznOVct2TStt/RyW+Z3NwOd8Bh0MA5TFEdv4wfkVBkd7d1ICvF9+hCz
+2LbVztZ4lkqjSooAFcz/xm1dXSNwhRwwG6sOAm4XC+7Pt6Oe5pq9ahFQKW+zkLr2
+TKQ8Ppr7wQW6IrrWola4MUeQ+DE8l4QKlLusvejPTdRsnnEt0RUPoWcACB1e0CzO
+3zOd300xy0ESeeHLLFuaG5FdOwIDAQABo1MwUTAdBgNVHQ4EFgQUhZdMK4poQSOG
+55sANxJlTMTCtEcwHwYDVR0jBBgwFoAUhZdMK4poQSOG55sANxJlTMTCtEcwDwYD
+VR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAgEAJNPRgCNHbSR+UYKFSet4
+XscYjtxb59QxFqN+AqUiEh38SVzWo2nhnDZxiDsTK2nA0qSrseoMTnTmWJuEf+io
+n3XmUNgalxzV0QGMhQZPT3hyoPditwc7easnm2Jwd10kkhMdlXnfy3lgHPIMH6/E
+0jh3sSuEora9Q3jIdobRQALrHI++lsLdOcJy0A2LKkVWbvftz6B2cDC6M+QhwPGo
+QNIuI8Cl2J89mNsDtZRohbA2/iUGE2Y47m9OD8snLOOZmhiCHsq82AYEt0wAqPI1
+biPvN/F+S4lMj71ue77BOg6sKKvotSgJQvyUO4iAouHoxxo5luT1nz2i2ofdvHFd
+Lm3eth9hEykOEi7AIJHSlXcFEdlEaP+ohOcTDoaoZc4kqYxuEDpC6otaQhHvqS/D
+0JGjdVTw7q0msfzG/7a5k4hklgB4JHrzWS++sCFKAu6LYnsVVFKu6P9qQxFv86qi
+xpWX9sj+7+qw1qPjSNKx9S6Lx+2vMIG/2l1rsJIKMLC4ndnT3bjwme0YS0/iOEi3
+EQT//Mb6Hg1R4r2OlvfYSt+pGOa64UwgqHj8wOLIkb/txsF3mQDbDE16tByih83E
+Wl44xDuoTuCZsvwNh9PJtg1XRaIdXwZFauRwV3nyLREeDX3EESmOyXYZZljcS9/K
+Gw265N/AwFB26FAC4Mof6FI=
+-----END CERTIFICATE-----
