+++ others/wildcard_cert.pem
@@ -1,0 +1,32 @@
+-----BEGIN CERTIFICATE-----
+MIIFlzCCA3+gAwIBAgIUJQM7MYyRjtvlVYzVqog0JGigp7swDQYJKoZIhvcNAQEL
+BQAwWjELMAkGA1UEBhMCQVUxEzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoM
+GEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDETMBEGA1UEAwwKKi5ub2RlLmNvbTAg
+Fw0yMjA5MjEwOTA3MDVaGA8yMDUwMDIwNjA5MDcwNVowWjELMAkGA1UEBhMCQVUx
+EzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMg
+UHR5IEx0ZDETMBEGA1UEAwwKKi5ub2RlLmNvbTCCAiIwDQYJKoZIhvcNAQEBBQAD
+ggIPADCCAgoCggIBAK62Sk9JUSQXv+wk5FztR1L0H8IgK34th5dOADV0zwY0Olnm
+ukkjI2cljNaY3DgfhGdv6m+3sozMliRjtmOkLtWoGsK8Bwv6lZrF6GPy2N0IeIQ5
+tddyaENm5ua7rG3kgYCBItadOg33A6eUWFYmZi5K7E9fpHUS7JCi+tYAaHqaESfI
+z2wAzlLo+nyfn6AzAWkcdtsNtwCJOXbC6gvJIjPwdsM9wJ77m1k32Dwo2ZJ4BAQm
+JoHNZaIy3pytZWaoIHnKwioIF/xKALls6A8TQqetPFxUfGQAsKqMFF/nbzn+S4GE
+ltufX/E1KECNiircHWC998XqLoaBim6d9+MKcyBZqG+Gr6k4JGLqw+FTVdCKo0hO
+L0GWA+TWZbDtCBmR2fEI8UJ3wnmDjzGir/9Yi/ugmovPQ5lpQIhEQWTQ/zGgVP38
+xVTYEPAdVVjFRnFhLlmRDedYo552A68KBlZPPotra5OustB6Rop9z4kNxH7mGDQg
+GdS6B0XbJUSk4325+7OQZpp30ursh+f4X0dVyC4maPjUC0YZuMbvoTLwwmu7zrtM
+NpvlLhcdBaBfKLdU6/bAxd7V8VHEXJOxnftPwmk151TxC3eHvHX85scNZU/nBl3R
+VyNgumcjsSzfb3z2NzHlBH5YKOmGWnDRsnMhMf7qO3k2YI0/XiV+Yd4jHGWvAgMB
+AAGjUzBRMB0GA1UdDgQWBBS76NBkLueNtmDSt+kF9y3Q75VuUDAfBgNVHSMEGDAW
+gBS76NBkLueNtmDSt+kF9y3Q75VuUDAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3
+DQEBCwUAA4ICAQAhkFEXcDVFZmHtAJ3KY5t2OOFBKV1JRU2tbcV11qTbXjCPASz7
+AejV+rVqHYDnHMo/lJ0l+OvDl/UDU8wIQo/eBqbba96ClvSrEnEi7Ekx5sfPrLBP
+Q1Mg3aKv9PMtOME/QAc4JExjHBz2ZGffM9EcF7bv4hIqeGOCmEV6yXSYj3PDuLL+
+3F8LF39eKjT5TldrhAkk+sAsSdLLu14oqlbUW61XRrqZaU4PprYdYOGEjUlWTrOH
+L60Qy8EsqGzzohU7wljm0U1m90rGZGYdaoRVNPwuvX8Olb72BQZU8OtUlJbFCHIE
+nPpHSN9sKtS1SHme3nYo/dOcuHhjXwoBXiWESEVCsJCAwubCKpj8YlB2JQqHKNUc
+fLQoueR6gBrsAP7cXAdEUnXX91cLYbetq09DXquV/b/YDctHuhRJO1kHrollQaWe
+XCFHpy3wJjxzoOMDrp1wsVidG8ogTkWxAwW1zbH/9tIW8cbvuVvoUXnzCO3CPBU6
+uOuOFLNFl2JJcF0Uddsg62LEOJkiVwhqmBFPWU9ywQ5lFrWK13uZjcnpeP7KWt3+
+0NqsRLX3uwU6yBd3aNcmps4yUPAjKbHCXm54/VBjuv1GOoBFWGrI0B5RjuZ+PZUH
+lYRmOVpgA3EioOvMukDE/p96Her3MISxplyU0loUlgXBbTj7VI212FQefg==
+-----END CERTIFICATE-----
