+++ others/wildcard_key.pem
@@ -1,0 +1,52 @@
+-----BEGIN PRIVATE KEY-----
+MIIJQQIBADANBgkqhkiG9w0BAQEFAASCCSswggknAgEAAoICAQCutkpPSVEkF7/s
+JORc7UdS9B/CICt+LYeXTgA1dM8GNDpZ5rpJIyNnJYzWmNw4H4Rnb+pvt7KMzJYk
+Y7ZjpC7VqBrCvAcL+pWaxehj8tjdCHiEObXXcmhDZubmu6xt5IGAgSLWnToN9wOn
+lFhWJmYuSuxPX6R1EuyQovrWAGh6mhEnyM9sAM5S6Pp8n5+gMwFpHHbbDbcAiTl2
+wuoLySIz8HbDPcCe+5tZN9g8KNmSeAQEJiaBzWWiMt6crWVmqCB5ysIqCBf8SgC5
+bOgPE0KnrTxcVHxkALCqjBRf5285/kuBhJbbn1/xNShAjYoq3B1gvffF6i6GgYpu
+nffjCnMgWahvhq+pOCRi6sPhU1XQiqNITi9BlgPk1mWw7QgZkdnxCPFCd8J5g48x
+oq//WIv7oJqLz0OZaUCIREFk0P8xoFT9/MVU2BDwHVVYxUZxYS5ZkQ3nWKOedgOv
+CgZWTz6La2uTrrLQekaKfc+JDcR+5hg0IBnUugdF2yVEpON9ufuzkGaad9Lq7Ifn
++F9HVcguJmj41AtGGbjG76Ey8MJru867TDab5S4XHQWgXyi3VOv2wMXe1fFRxFyT
+sZ37T8JpNedU8Qt3h7x1/ObHDWVP5wZd0VcjYLpnI7Es32989jcx5QR+WCjphlpw
+0bJzITH+6jt5NmCNP14lfmHeIxxlrwIDAQABAoICAA8++xcWJ82FgQsDYSY20o2L
+niXbltAI+xcaCGyhx9sbvafQuZreRR2gKj0j07vWDEDWhFfBaQReag+839KsJiYg
+6EzsCZVjBaEe+huWupP3REUNoC+v3SdcmHO0FuT0FtT27+pYiBNl5dy+1kKWUT5i
+7WIzz28j+p7YihkYqgVg4nRdUrRzEY4spIcxisV5Dft1L24IMVsLnocdlTz4DVMI
+/eQlxGRbQoAznlwJmIrujDMsfRnAqcjQtZpoizu9GjBmoDD4ydcLQirsNQfzv1JQ
+jL7l5ID5inEnjjOcR5mA9mfUATIjI0UyRVP0xgTddnbVvQR5XhwAu+rRQCuA4vDf
+0oANv5EEzd8DISViISxV1Z2UUYjBSEpcW/y/BFQxYOsxvjqQnySyDCj6rzvWxHXX
+4PHGKQ17J2CnRWljcI3xMZkL5K2gzabsnmxOPZeBgncL7UBLn1hMzkDpU94zfr8t
+TnFXWF3I1zB5nYNUfXNNf9EYpoi2mJ9Wvgd5b49Tkd1QF4/9huf3PP0sBzqGhC0Z
+p+DksusOh0PjgiYFE89np5wchnOSQf8ySmGw5QaKiOw0hlGti1AQbsjnWTIR00Py
+1DhA3IRhS/Iocnw+AknUqgqzV7tCSE0jdO0iBNAA2QE/M6/fvnfI8g6mnWv4AFI0
+1K+Y+fOOLgA+meflGCPJAoIBAQDL0Tqj9bs0lxoRtDx0MdYxhSATUzcup4eFpUt2
+5BsqM5nWDon6+n/xta5TcQ7hBYHZEZv3or0vINH93XZ+JYOmTrc6V464OGUyOlzl
+WnLcN0DGAjBD1aQ37aPxeNZ8+8lCsd7zgE6xaGDx618fYBFDGpTHpaxURSh2v8dl
+sjM14YsV/Kw5rlo6n76PGcVafgLk0acwJXQFfsg/KdQqhL0N+o+f8pFrEsGfV7QU
+JiaXvljzN8VfPmAIblzB/IwSEj+pEhGQHqxuMDSpZS61W+hWTRFFaBRy5MBqHTY0
+A9zfPJwCZNdfvQ5tCc4gjRBFdK/7KklXjTwvvz//s60iAhHrAoIBAQDbcWvBCWUF
+tJhB7xQM/24x0WorO21tiTurVh/D5ASYZEoZLDU0O7fN4iaoLjgh7MbYpNb+fY7p
+d5cuK63Tot9ckZF96KuDedYQPkK0diT3YYS+f6+zdgWEHZOB+sl8LwhYXs7WbSQK
+oD2nYFbIn0GXkh6hepZf4RgBGHE6wmxwzP3pEtze5ENta+GQ0vqhJncXpifZth03
+WeRsbLzbMpqxogvilTyf5/hqZ8e1jw6h5Y88JCLeupJPhWjAWzoyH8/ATZUoOG6g
+rEkCAjaantSzJv85PLIAEqT9b5bsgiPjMAdCxxXI0qcnvFqA+kxHYba7jLbBtHN6
+Ww+5ZyFWVIZNAoIBAHWXMmD0Xly6LOU3+oUDV9PcPaRIp5a4Oz1FH6PpeS7A9KmG
+W988dWLL8lbb7Lyww7isKzMVZlbxdZYlFr8y/6CM27PCXmo0x1XSTFRa+dfJ3Qog
+qx4H/aXy6lpf4EfdsMFVBA+y4DQEx6HrrUbZ0zvG8z24fRNljzHr9avbwyX8CZXW
+4LqquA4IqhW/xZgvCZZgyzfA9A5Tlk6y4GDjknulKwYaGgIPoTxAinnxhOSZgpJY
+g/wubZOPsN6OyXWHp8jwzsc6mzs7fraY1Pj3nLVl42gK43liNw1B60rZS65UNqI+
+8e+fkNoiBRXMrS2VZA7h18y2hJn7mn35eRMc5J8CggEAGml8Xq+SyS0DZmwwLaBQ
+0w4xiSKN+TEbTXQAPsOjCfSINp9rh+3oIY+mt3dYqjylck3k0pyqJs+OErCPK3nf
+CHU83bqzag+wBCR6Qt41dND3+TXFIASEhXJJ3ssglSHUD07Z9lUMkXrX8N1XlK4W
+ZvffV/A5STD+gqM+faQJCUqjBPqQOwOeWf0yGQxLvjmy+9Wz1jBlM9RN+f7cfEWK
++PZYF91+foCdFbGICAI/8JyAsOxohgZOteOIyB4y0vWhO7Qaz8x7BDu4TdWWXOXh
+E+8NqQX1Jq0XzqkvsSe4yg9t0EvGA3XXSKETA5Mcuzq3k7tMA+SEQaXULqyLBqkF
+wQKCAQAIDGJOIIhejb8H3hKnGi3bIxICU6wItGWA9a4d5TLT7Qf1BgC7n0FNtYeU
+C6Z98s1LG+iu8vTk2q+vnhMH+SBkMhWzPGzgNThd7lNJKUuY9/wArTf/uyQLSkyT
+0au+LD+NWCqgF59UUz/t3UfgEASaMW+WGqBPLlYr5suwdYpQ+WsKFEIapFlKBUfE
+Kj/+Kl6Ejq5QL5d5D7076QOiH5zxydRYPgck/OL4Kk43DlPuBxk1lShUh0S/UNpK
+qJi5PYglxtpI3m3dydugeo408pgEWHf6izoay09Cl34k6bH8Gz+oMX60oJmcROp1
+SOS2HtRM0ZdpfQqThSDEKSQ0xkt6
+-----END PRIVATE KEY-----
